Fix SingleUserPicker filtering
