+++ src/modules/core/components/SingleUserPicker/index.ts
@@ -1,3 +1,5 @@
 export { default } from './SingleUserPicker';
 
 export { default as ItemDefault } from './ItemDefault';
+
+export * from './utils';
