+++ src/modules/users/transformers.ts
@@ -1,7 +1,9 @@
 import { UsersMapType } from './state/index';
 
 export const getUserPickerData = (userData: UsersMapType) =>
-  Object.entries(userData).map(([address, { record }]) => ({
-    id: address,
-    ...record,
-  }));
+  Object.entries(userData)
+    .filter(([, { record }]) => !!(record && record.profile))
+    .map(([address, { record }]) => ({
+      id: address,
+      ...record,
+    }));
