+++ src/utils/arrays/index.ts
@@ -200,35 +200,3 @@ export const recursiveNestChildren = (
   });
   return collapsedItems;
 };
-
-/*
-Extracts the required values to be used in the SingleUserPicker
-on selection
-*/
-export const filterUserSelection = (data, filterValue) => {
-  if (!filterValue) {
-    return data;
-  }
-
-  const filtered = data.filter(
-    user =>
-      user &&
-      filterValue &&
-      (user.profile.username
-        .toLowerCase()
-        .includes(filterValue.toLowerCase()) ||
-        user.profile.walletAddress
-          .toLowerCase()
-          .includes(filterValue.toLowerCase())),
-  );
-
-  const customValue = {
-    id: 'filterValue',
-    profile: {
-      walletAddress: filterValue,
-      displayName: filterValue,
-    },
-  };
-
-  return [customValue].concat(filtered);
-};
